% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperMethods.R
\name{applyMapper}
\alias{applyMapper}
\title{applyMapper}
\usage{
applyMapper(data, mapper)
}
\arguments{
\item{data}{A data matrix.}

\item{mapper}{Existing mapper network}
}
\description{
Map new data to existing Mapper network.
}
