% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperMethods.R
\name{detect_communities}
\alias{detect_communities}
\title{Perform community detection on the existing graph
 with a new resolution, then re-label the communities.}
\usage{
detect_communities(mapper, data, resolution)
}
\arguments{
\item{mapper}{Existing mapper object.}

\item{data}{The data matrix used for obtaining mapper object.}

\item{resolution}{Numeric, resolution for the Leiden algorithm.}
}
\description{
Perform community detection on the existing graph
 with a new resolution, then re-label the communities.
}
