% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperMethods.R
\name{extract_features}
\alias{extract_features}
\title{extract_features}
\usage{
extract_features(data, sample_mapping, mapper)
}
\arguments{
\item{data}{A data matrix or data frame, with observations
along rows and variables along columns.}

\item{sample_mapping}{A vector giving the sample membership of
each data point.}

\item{mapper}{Existing mapper object.}
}
\description{
Extract features (cell type proportions across samples)
node compositions (node proportions across samples).
}
