% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperMethods.R
\name{flowMapper}
\alias{flowMapper}
\title{flowMapper}
\usage{
flowMapper(
  data,
  nx = 10,
  ny = 10,
  overlap = 0.1,
  nodemax = 1000,
  intersectionSize = 10
)
}
\arguments{
\item{data}{A data matrix.}

\item{nx}{Number of bins (level sets) for the first
filter dimension.}

\item{ny}{Number of bins (level sets) for the second
filter dimension.}

\item{overlap}{Fractional overlap between neighboring bins.}

\item{nodemax}{In level set clustering stage, split level
sets with larger cardinality.}

\item{intersectionSize}{Minimum number of shared datapoints
required for an edge.}
}
\description{
Wrapper for the core Mapper functionality.
}
