% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperMethods.R
\name{getCommunitiesFeatures}
\alias{getCommunitiesFeatures}
\title{getCommunitiesFeatures}
\usage{
getCommunitiesFeatures(
  data,
  sample_mapping,
  mapper,
  defs,
  additional = list(),
  resolution = 2
)
}
\arguments{
\item{mapper}{Existing mapper object.}
}
\description{
Wrapper for communitiy detection, labeling,
and extracting features (cell type percentages).
}
