% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperMethods.R
\name{getFilter}
\alias{getFilter}
\title{getFilter}
\usage{
getFilter(data, scale = FALSE, plotPath = NULL)
}
\arguments{
\item{data}{A data matrix or data frame, with observations
along rows and variables along columns.}

\item{scale}{Logical, whether to scale the data before PCA.}

\item{plotPath}{Generate plots of the filter color-coded by
each of the variables in the data matrix and save them at this path.
No plots generated unless path is provided.}
}
\description{
Get the filter function (PCA projection) used by HiTMapper.
}
