% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperMethods.R
\name{getSampleFeatures}
\alias{getSampleFeatures}
\title{getSampleFeatures}
\usage{
getSampleFeatures(nodes, sampleMapping, cl, normalized = TRUE)
}
\arguments{
\item{nodes}{A list of integer vectors, each containing the data
points assigned to each node.}

\item{sampleMapping}{An integer vector, containing the sample of
origin for each data point.}

\item{normalized}{Logical, whether the contingency table should
be normalized across features. (Normalization across samples takes
place either way.}
}
\description{
Create a contingency table of sample vs node membership.
}
