% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperMethods.R
\name{getSampleFeaturesSmooth}
\alias{getSampleFeaturesSmooth}
\title{Moving average features.}
\usage{
getSampleFeaturesSmooth(mapper, sampleMapping, norm = FALSE)
}
\arguments{
\item{sampleMapping}{An integer vector, containing the sample of
origin for each data point.}

\item{nodes}{A list of integer vectors, each containing the data
points assigned to each node.}

\item{normalized}{Logical, whether the contingency table should
be normalized across features. (Normalization across samples takes
place either way.}
}
\description{
Extract features from multiple samples, as a contingency table
of sample vs node membership. Smoothed over neighboring nodes in the network
for de-noising purposes, analogous to a moving average.
}
