% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperMethods.R
\name{HiTMapper}
\alias{HiTMapper}
\title{HiTMapper}
\usage{
HiTMapper(
  data,
  scale = FALSE,
  nx = 10,
  ny = 10,
  overlap = 0.3,
  nodemax = NULL,
  kmax = NULL,
  kNodes = 500,
  intersectionSize = NULL,
  iou = 0.03,
  clusterMethod = "kmeans",
  outlierCutoff = 25,
  verbose = FALSE
)
}
\arguments{
\item{data}{A data matrix.}

\item{nx}{Number of bins (level sets) for the first
filter dimension.}

\item{ny}{Number of bins (level sets) for the second
filter dimension.}

\item{overlap}{Fractional overlap between neighboring bins.}

\item{nodemax}{In level set clustering stage, split level
sets with larger cardinality.}

\item{intersectionSize}{Minimum number of shared datapoints
required for an edge.}
}
\description{
Wrapper for the core Mapper functionality.
}
