% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperMethods.R
\name{HiTMapper_predict}
\alias{HiTMapper_predict}
\title{Use existing mapper model to predict cluster membership of new data.}
\usage{
HiTMapper_predict(mapper, data)
}
\arguments{
\item{mapper}{Existing mapper object.}

\item{data}{New data matrix, must have same columns as that used for
mapper model.}
}
\description{
Use existing mapper model to predict cluster membership of new data.
}
