% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperMethods.R
\name{HiTMapperPooled}
\alias{HiTMapperPooled}
\title{HiTMapperPooled}
\usage{
HiTMapperPooled(
  path,
  files,
  ignore = c(),
  kNodes = 500,
  nx = 7,
  ny = 7,
  overlap = 0.2
)
}
\arguments{
\item{path}{The path where fcs files are.}

\item{files}{List of files to use.}

\item{ignore}{Optional, list of markers to ignore.}
}
\description{
Pool data from multiple fcs files, and compute a
joint model using a single layout.
}
