% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperMethods.R
\name{HiTMapperPooledMatrix}
\alias{HiTMapperPooledMatrix}
\title{HiTMapperPooledMatrix}
\usage{
HiTMapperPooledMatrix(
  data,
  sample_mapping,
  kNodes = 500,
  nx = 7,
  ny = 7,
  overlap = 0.2
)
}
\arguments{
\item{data}{A data matrix.}

\item{sample_mapping}{A character vector assigning cells to samples.}
}
\description{
Take as input a data matrix and a character vector assigning
cells to samples, then compute a joint model using a single layout.
}
