% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperMethods.R
\name{HiTMapperSamplewise}
\alias{HiTMapperSamplewise}
\title{HiTMapperSamplewise}
\usage{
HiTMapperSamplewise(
  path,
  files,
  ignore = c(),
  kNodes = 400,
  nx = 7,
  ny = 7,
  overlap = 0.2
)
}
\arguments{
\item{path}{The path where fcs files are.}

\item{files}{List of files to use.}

\item{ignore}{Optional, list of markers to ignore.}
}
\description{
Fit HiTMapper models independently to each sample,
then align them to a common layout.
}
