% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperMethods.R
\name{label_communities}
\alias{label_communities}
\title{label_communities}
\usage{
label_communities(mapper, defs, additional = list())
}
\arguments{
\item{mapper}{Existing mapper object.}

\item{defs}{Data frame of phenotype definitions.}

\item{additional}{Named list, specifying attributes to be
appended to phenotypes. E.g. list(CD38="activated", Ki67="proliferating").}
}
\description{
Wrapper for community detection, labeling,
and extracting features (cell type percentages).
}
