% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperMethods.R
\name{louvainClustering}
\alias{louvainClustering}
\title{louvainClustering}
\usage{
louvainClustering(graph, gamma = 1)
}
\arguments{
\item{graph}{An igraph object, for example, from the output of flowMapper.}

\item{gamma}{Numeric value controlling the number of meta-clusters:
increase for more, decrease for fewer.}
}
\description{
Meta-clustering of Mapper nodes using Louvain method.
}
