% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lymphocytes.R
\docType{data}
\name{lymphocytes}
\alias{lymphocytes}
\title{Flow cytometry measurements for 19225 cells.}
\format{
A data frame with 19225 rows and 11 variables:
\describe{
  \item{GFP}{Green fluorescent protein}
  \item{CD8}{Lineage marker for cytotoxic T cells}
  \item{l.d}{Live/Dead marker; data is pre-gated on live cells}
  \item{UNK}{Unknown marker with bimodal distribution}
  \item{TCRyd}{T cell receptor gamma-delta}
  \item{CD45}{Lineage marker for all lymphocytes}
  \item{TCRb}{T cell receptor alpha-beta}
  \item{NK1.1}{Lineage marker for natural killer cells}
  \item{CD4}{Lineage marker for helper T cells}
  \item{CD19}{Lineage marker for B cells}
  \item{CD3}{Lineage marker for T cells}
}
}
\usage{
lymphocytes
}
\description{
A dataset containing compensated, logicle-transformed
values of 11-color flow cytometry data.
}
\keyword{datasets}
