% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperMethods.R
\name{nodeComposition}
\alias{nodeComposition}
\title{nodeComposition}
\usage{
nodeComposition(mapper, samples, scale = FALSE)
}
\arguments{
\item{mapper}{Existing mapper object.}

\item{samples}{An integer vector, containing the sample of
origin for each data point.}

\item{scale}{Logical, whether the contingency table should
be scaled across features. (Scaling across samples takes
place either way.}
}
\description{
Contingency table of sample vs node membership.
}
