% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperMethods.R
\name{plot_mapper}
\alias{plot_mapper}
\title{plot_mapper}
\usage{
plot_mapper(
  mapper,
  markers = colnames(mapper$centroids),
  path = NULL,
  device = "png"
)
}
\arguments{
\item{mapper}{Existing mapper object.}

\item{markers}{A subset of the marker names.}

\item{path}{The path where plots are saved.}

\item{device}{The device used for image encoding. Supports
png for bitmaps, pdf for vector graphics.}
}
\description{
Generates plots of the network color-coded by
each of the specified markers.
}
