% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingMethods.R
\name{plot_mapper_interactive}
\alias{plot_mapper_interactive}
\title{plot_mapper_interactive}
\usage{
plot_mapper_interactive(
  mapper,
  color = mapper$community,
  labels = paste0("Community: ", mapper$community, "\\n"),
  path = "",
  title = "HiTMapper"
)
}
\arguments{
\item{mapper}{Existing mapper object.}

\item{color}{A factor or numeric vector which is mapped to the
color of the nodes.}

\item{labels}{A character vector describing each node, displayed
when hovering over the node.}

\item{path}{The path where plots are saved.}

\item{title}{The name of the output html file.}
}
\description{
Generates an interactive plot of the network and
saves it as an html file.
}
