% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperMethods.R
\name{pruneEdges}
\alias{pruneEdges}
\title{pruneEdges}
\usage{
pruneEdges(mapper, cutoff = 0.01)
}
\arguments{
\item{mapper}{Existing mapper object.}

\item{cutoff}{Prune edges whose weight is smaller than this fraction of
the maximum weight.}
}
\description{
Prune edges with low weight, to obtain a sparser network
and minimize spurious connections.
}
