% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{Flow cytometry measurements for 10,000 cells.}
\format{
A data frame with 10,000 rows and 26 variables:
\describe{
  \item{CD8}{Lineage marker for cytotoxic T cells}
  \item{TCRgd}{T cell receptor gamma-delta}
  \item{CD45}{Lineage marker for lymphocytes}
  \item{CD4}{Lineage marker for helper T cells}
  \item{CD19}{Lineage marker for B cells}
  \item{CD3}{Lineage marker for T cells}
}
}
\usage{
toy_data
}
\description{
A dataset containing compensated, biexponentially transformed
values of 26-color flow cytometry data.
}
\keyword{datasets}
