% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperMethods.R
\name{detect_communities}
\alias{detect_communities}
\title{detect_communities}
\usage{
detect_communities(mapper, data, resolution = 8, merge = TRUE)
}
\arguments{
\item{mapper}{Existing mapper object.}

\item{data}{A data matrix or data frame, with observations
along rows and variables along columns.}

\item{resolution}{passed to the Leiden algorithm, controls
the number of communities. Increase for more, decrease for fewer.}

\item{merge}{Logical, whether to merge communities which have the same
modality across all markers in the data. It is recommended to use a
large resolution to obtain many communities, some of which are then merged.}
}
\description{
Use Leiden algorithm to identify communities in the
HiTMapper network. A community is a group of similar nodes,
which approximates the concept of "cell type".
}
