% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperMethods.R
\name{HiTMapper}
\alias{HiTMapper}
\title{HiTMapper}
\usage{
HiTMapper(
  data,
  total_nodes,
  method = "som",
  min_node_size = 50,
  grid_size = c(10, 10),
  overlap = 0,
  n_passes = 10,
  scale = FALSE,
  filter = NULL,
  resolution = 8,
  defs = NULL,
  k = 8,
  force_sep = c()
)
}
\arguments{
\item{data}{A data matrix or data frame, with observations
along rows and variables along columns.}

\item{total_nodes}{Approximate number of nodes for the Mapper graph.}

\item{method}{Clustering method to use for each bin. Currently implemented
options are "som" (recommended) or "kmeans".}

\item{min_node_size}{Minimum number of cells in a node. No need to change
the default, unless there are very few total cells.}

\item{grid_size}{A vector of integers, specifying the number of
overlapping bins along each principal component. Its length determines
the number of principal components used.}

\item{overlap}{Fraction of overlap between neighboring bins.}

\item{n_passes}{Number of passes over all data points when allocating
cells to nodes via the "som" method.}

\item{scale}{Logical, whether to scale the data before PCA.}

\item{filter}{A matrix with few columns and same number of rows as data.
To be used as filter, instead of the default principal components.}

\item{resolution}{Used in community detection. Larger values lead to more
communities.}

\item{k}{Number of nearest neighbors to use when constructing graph.}

\item{force_sep}{Character vector, must be a subset of colnames(data).
Weakens graph edges between nodes that have different modality for any
of the given markers. This is a kludge and should be used sparingly.}
}
\description{
Wrapper for the core Mapper functionality.
}
