% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperMethods.R
\name{HiTMapper}
\alias{HiTMapper}
\title{HiTMapper}
\usage{
HiTMapper(
  data,
  total_nodes,
  method = "som",
  graph = "knn",
  outlier_cutoff = nrow(data)/10000,
  grid_size = c(10, 10),
  overlap = 0.15,
  n_passes = 10,
  scale = FALSE,
  verbose = FALSE,
  filter = NULL,
  merge = TRUE,
  split = c(),
  resolution = 2,
  k = 8
)
}
\arguments{
\item{data}{A data matrix or data frame, with observations
along rows and variables along columns.}

\item{total_nodes}{Approximate number of nodes for the Mapper graph.}

\item{outlier_cutoff}{Discard nodes containing fewer cells than this.}

\item{grid_size}{A vector of integers, specifying the number of
overlapping bins along each principal component. Its length determines
the number of principal components used.}

\item{overlap}{Fraction of overlap between neighboring bins.}

\item{scale}{Logical, whether to scale the data before PCA.}

\item{verbose}{Logical, whether to output detailed info.}
}
\description{
Wrapper for the core Mapper functionality.
}
