% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperMethods.R
\name{HiTMapper}
\alias{HiTMapper}
\title{HiTMapper}
\usage{
HiTMapper(
  data,
  total_nodes = 1000,
  resolution = 1,
  defs = NULL,
  grid_size = c(5, 4),
  min_node_size = 50,
  n_passes = 10
)
}
\arguments{
\item{data}{A data matrix, with observations
along rows and variables along columns.}

\item{total_nodes}{Approximate number of nodes for the Mapper graph.}

\item{resolution}{Used in community detection. Larger values lead to more
communities. We recommend running HiTMapper with default resolution,
and later customizing if necessary using the detect_communities function,
at very little computational cost.}

\item{defs}{A data frame of cell population definitions, to be used for
labeling communities.}

\item{grid_size}{Array of integers, the number of bins along 
each principal component.}

\item{min_node_size}{Minimum number of cells in a node.}

\item{n_passes}{Number of passes over all data points when clustering.}
}
\description{
Wrapper for the core Mapper functionality.
}
