% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperMethods.R
\name{label_communities}
\alias{label_communities}
\title{label_communities}
\usage{
label_communities(mapper, defs, thresholds = NULL)
}
\arguments{
\item{mapper}{Existing mapper object.}

\item{defs}{Data frame of phenotype definitions.}

\item{thresholds}{Data frame of user-supplied thresholds,
for a subset of markers,
to override those estimated by the algorithm.}
}
\description{
Wrapper for community detection and labeling.
}
