% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperMethods.R
\name{leidenClustering}
\alias{leidenClustering}
\title{leidenClustering}
\usage{
leidenClustering(graph, resolution = 1)
}
\arguments{
\item{graph}{An igraph object, for example, from the output of HiTMapper.}

\item{resolution}{Numeric value controlling the number of communities:
increase for more, decrease for fewer.}
}
\description{
Community detection using Leiden method.
}
